3.4.10. Adpated to pvr addon api 5.2.1 (3.4.9 was just to ake it compile again).

+++ src/Tvheadend.cpp
@@ -2466,7 +2466,7 @@ void CTvheadend::DemuxAbort ( void )
   }
 }
 
-bool CTvheadend::DemuxSeek ( int time, bool backward, double *startpts )
+bool CTvheadend::DemuxSeek ( double time, bool backward, double *startpts )
 {
   return m_dmx_active->Seek(time, backward, startpts);
 }
